pyproject.toml
[project]
name = "testwork"
version = "0.1.0"
description = ""
authors = [
    {name = "Berry_Hunter",email = "tank-20@list.ru"}
]
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "django (>=6.0.4,<7.0.0)",
    "djangorestframework (>=3.17.1,<4.0.0)",
    "psycopg2 (>=2.9.12,<3.0.0)",
    "djangorestframework-simplejwt (>=5.5.1,<6.0.0)",
    "load-dotenv (>=0.1.0,<0.2.0)",
    "django-filter (>=25.2,<26.0)",
]


[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"
